% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_angry.R
\name{make_angry}
\alias{make_angry}
\title{Make your string angry}
\usage{
make_angry(string)
}
\arguments{
\item{string}{a string to make angry.}
}
\value{
A string.
}
\description{
Make your string angry
}
\examples{
x <- "hello"
make_angry(x)
}
