% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sad.R
\name{make_sad}
\alias{make_sad}
\title{Make your string sad}
\usage{
make_sad(string)
}
\arguments{
\item{string}{a string to make sad}
}
\value{
A string.
}
\description{
Make your string sad
}
\examples{
x <- "Hello"
make_sad(x)
}
